test: ITStorage converted to ITObjectTest
